COL-1776, edit title and users in context of /whiteboard canvas
