+++ src/components/whiteboards/toolbar/assets/UploadNewAsset.vue
